Consolidate Block Structure folders
TNL-6518
