+++ lms/djangoapps/course_api/blocks/transformers/block_depth.py
@@ -1,7 +1,7 @@
 """
 Block Depth Transformer
 """
-from openedx.core.lib.block_structure.transformer import BlockStructureTransformer
+from openedx.core.djangoapps.content.block_structure.transformer import BlockStructureTransformer
 
 
 class BlockDepthTransformer(BlockStructureTransformer):
