+++ lms/djangoapps/course_api/blocks/transformers/blocks_api.py
@@ -1,7 +1,7 @@
 """
 Blocks API Transformer
 """
-from openedx.core.lib.block_structure.transformer import BlockStructureTransformer
+from openedx.core.djangoapps.content.block_structure.transformer import BlockStructureTransformer
 from .block_counts import BlockCountsTransformer
 from .block_depth import BlockDepthTransformer
 from .navigation import BlockNavigationTransformer
