+++ lms/djangoapps/course_api/blocks/transformers/navigation.py
@@ -1,7 +1,7 @@
 """
 TODO
 """
-from openedx.core.lib.block_structure.transformer import BlockStructureTransformer
+from openedx.core.djangoapps.content.block_structure.transformer import BlockStructureTransformer
 from .block_depth import BlockDepthTransformer
