+++ lms/djangoapps/course_api/blocks/transformers/student_view.py
@@ -1,7 +1,7 @@
 """
 Student View Transformer
 """
-from openedx.core.lib.block_structure.transformer import BlockStructureTransformer
+from openedx.core.djangoapps.content.block_structure.transformer import BlockStructureTransformer
 
 
 class StudentViewTransformer(BlockStructureTransformer):
