+++ lms/djangoapps/course_api/blocks/transformers/tests/test_block_depth.py
@@ -7,8 +7,8 @@ Tests for BlockDepthTransformer.
 import ddt
 from unittest import TestCase
 
-from openedx.core.lib.block_structure.tests.helpers import ChildrenMapTestMixin
-from openedx.core.lib.block_structure.block_structure import BlockStructureModulestoreData
+from openedx.core.djangoapps.content.block_structure.tests.helpers import ChildrenMapTestMixin
+from openedx.core.djangoapps.content.block_structure.block_structure import BlockStructureModulestoreData
 from ..block_depth import BlockDepthTransformer
