+++ lms/djangoapps/course_api/blocks/transformers/tests/test_student_view.py
@@ -4,7 +4,7 @@ Tests for StudentViewTransformer.
 
 # pylint: disable=protected-access
 
-from openedx.core.lib.block_structure.factory import BlockStructureFactory
+from openedx.core.djangoapps.content.block_structure.factory import BlockStructureFactory
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import ToyCourseFactory
