+++ lms/djangoapps/course_blocks/api.py
@@ -3,7 +3,7 @@ API entry point to the course_blocks app with top-level
 get_course_blocks function.
 """
 from openedx.core.djangoapps.content.block_structure.api import get_block_structure_manager
-from openedx.core.lib.block_structure.transformers import BlockStructureTransformers
+from openedx.core.djangoapps.content.block_structure.transformers import BlockStructureTransformers
 
 from .transformers import (
     library_content,
