+++ openedx/core/djangoapps/content/block_structure/block_structure.py
