+++ openedx/core/djangoapps/content/block_structure/tests/test_transformers.py
@@ -57,7 +57,7 @@ class TestBlockStructureTransformers(ChildrenMapTestMixin, TestCase):
     def test_collect(self):
         with mock_registered_transformers(self.registered_transformers):
             with patch(
-                'openedx.core.lib.block_structure.tests.helpers.MockTransformer.collect'
+                'openedx.core.djangoapps.content.block_structure.tests.helpers.MockTransformer.collect'
             ) as mock_collect_call:
                 BlockStructureTransformers.collect(block_structure=MagicMock())
                 self.assertTrue(mock_collect_call.called)
@@ -66,7 +66,7 @@ class TestBlockStructureTransformers(ChildrenMapTestMixin, TestCase):
         self.add_mock_transformer()
 
         with patch(
-            'openedx.core.lib.block_structure.tests.helpers.MockTransformer.transform'
+            'openedx.core.djangoapps.content.block_structure.tests.helpers.MockTransformer.transform'
         ) as mock_transform_call:
             self.transformers.transform(block_structure=MagicMock())
             self.assertTrue(mock_transform_call.called)
