+++ openedx/core/djangoapps/content/block_structure/transformer.py
